requirements.txt
Django
djangorestframework
gunicorn
celery
